requirements.txt
numpy
skyfield
scipy
xarray
rioxarray
affine
netcdf4
pyproj
python-dateutil
Pillow
aggdraw
pyshp
